fix: Resolve Supabase connection and ensure Google Auth is stable
